Functional for msg embeds

+++ antifred.py
@@ -23,7 +23,7 @@ async def clearfred(ctx: interactions.SlashContext):
 
 @bot.event
 async def on_message_create(message):
-    if message.author.id == 184405311681986560:
+    if message.author.id == 184405311681986560 and message.embeds:
         await asyncio.sleep(10)
         try:
             await message.delete()
